regorg packages to be more human readable and cleaner
