+++ cmd/initialize/main_test.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/calypr/git-drs/internal/common"
 	"github.com/calypr/git-drs/internal/drslog"
+	"github.com/calypr/git-drs/internal/drspaths"
 	"github.com/calypr/git-drs/internal/gitrepo"
 	"github.com/calypr/git-drs/internal/testutils"
 )
@@ -102,8 +102,8 @@ func TestEnsureInitialized(t *testing.T) {
 		t.Fatalf("EnsureInitialized second call error: %v", err)
 	}
 
-	if _, err := os.Stat(common.DRS_DIR); err != nil {
-		t.Fatalf("expected %s to exist: %v", common.DRS_DIR, err)
+	if _, err := os.Stat(drspaths.DRSDir); err != nil {
+		t.Fatalf("expected %s to exist: %v", drspaths.DRSDir, err)
 	}
 	filterProcess, err := gitrepo.GetGitConfigString("filter.drs.process")
 	if err != nil {
