+++ cmd/pull/pull_test.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/calypr/git-drs/internal/config"
 	"github.com/calypr/git-drs/internal/lfs"
+	"github.com/calypr/git-drs/internal/remoteruntime"
 )
 
 func resetPullFlagsForTest() {
@@ -48,8 +49,8 @@ func TestPullDryRunListsMatchingPaths(t *testing.T) {
 
 	loadCfg = func() (*config.Config, error) { return &config.Config{}, nil }
 	resolveRemote = func(cfg *config.Config, name string) (config.Remote, error) { return config.Remote("origin"), nil }
-	newRemoteClient = func(cfg *config.Config, remote config.Remote, logger *slog.Logger) (*config.GitContext, error) {
-		return &config.GitContext{}, nil
+	newRemoteClient = func(cfg *config.Config, remote config.Remote, logger *slog.Logger) (*remoteruntime.GitContext, error) {
+		return &remoteruntime.GitContext{}, nil
 	}
 	loadWorktreeInventory = func(_ *slog.Logger) (map[string]lfs.LfsFileInfo, error) {
 		return map[string]lfs.LfsFileInfo{
