--- internal/common/common.go
@@ -1,44 +1,0 @@
-package common
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/bytedance/sonic"
-	drsapi "github.com/calypr/syfon/apigen/client/drs"
-)
-
-// ParseOrgProject resolves the effective org and project from the conventional
-// arguments. When org is provided it is used directly. When org is empty the
-// project string is split on the first "-" to derive org and project.
-func ParseOrgProject(org, project string) (string, string) {
-	if org != "" {
-		return org, project
-	}
-	if project == "" {
-		return "", ""
-	}
-	if !strings.Contains(project, "-") {
-		return "default", project
-	}
-	parts := strings.SplitN(project, "-", 2)
-	return parts[0], parts[1]
-}
-
-// PrintDRSObject marshals and prints a DRS object as JSON.
-func PrintDRSObject(obj drsapi.DrsObject, pretty bool) error {
-	var out []byte
-	var err error
-
-	if pretty {
-		out, err = sonic.ConfigFastest.MarshalIndent(obj, "", "  ")
-	} else {
-		out, err = sonic.ConfigFastest.Marshal(obj)
-	}
-	if err != nil {
-		return err
-	}
-
-	fmt.Printf("%s\n", string(out))
-	return nil
-}
