--- internal/common/values.go
@@ -1,12 +1,0 @@
-package common
-
-const (
-	LFS_OBJS_PATH string = ".git/lfs/objects"
-	DRS_OBJS_PATH string = ".git/drs/lfs/objects"
-	CONFIG_YAML   string = "config.yaml"
-
-	DRS_REF_DIR     string = ".git/drs/objects"
-	DRS_LOG_FILE    string = ".git/drs/drs.log"
-	ConfirmationYes string = "yes"
-	DRS_DIR         string = ".git/drs"
-)
