+++ internal/config/remote.go
@@ -1,43 +1,11 @@
 package config
 
-import (
-	"context"
-	"fmt"
-	"log/slog"
-	"net/url"
-	"strings"
-	"time"
-
-	calyprconf "github.com/calypr/calypr-cli/conf"
-	"github.com/calypr/calypr-cli/credentials"
-	"github.com/calypr/git-drs/internal/gitrepo"
-	syclient "github.com/calypr/syfon/client"
-	syconf "github.com/calypr/syfon/client/config"
-)
-
-const credentialHelpSuffix = "Refresh credentials with `git drs remote add gen3 <remote-name> <organization/project> --cred <path>` or `--token <token>`. See docs/getting-started.md."
-
 type DRSRemote interface {
 	GetProjectId() string
 	GetOrganization() string
 	GetEndpoint() string
 	GetBucketName() string
 	GetStoragePrefix() string
-	GetClient(remoteName string, logger *slog.Logger) (*GitContext, error)
-}
-
-type GitContext struct {
-	Client             *syclient.Client
-	Organization       string
-	ProjectId          string
-	BucketName         string
-	StoragePrefix      string
-	Upsert             bool
-	ForceUpload        bool
-	MultiPartThreshold int64
-	UploadConcurrency  int
-	Logger             *slog.Logger
-	Credential         *syconf.Credential
 }
 
 type RemoteSelect struct {
@@ -58,21 +26,6 @@ func (s Gen3Remote) GetOrganization() string  { return s.Organization }
 func (s Gen3Remote) GetEndpoint() string      { return s.Endpoint }
 func (s Gen3Remote) GetBucketName() string    { return s.Bucket }
 func (s Gen3Remote) GetStoragePrefix() string { return s.StoragePrefix }
-
-func (s Gen3Remote) GetClient(remoteName string, logger *slog.Logger) (*GitContext, error) {
-	manager := calyprconf.NewConfigure(logger)
-	cred, err := manager.Load(remoteName)
-	if err != nil {
-		return nil, err
-	}
-	ctx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-	defer cancel()
-	if err := credentials.EnsureValidCredential(ctx, cred, logger); err != nil {
-		return nil, WrapCredentialValidationError(remoteName, err)
-	}
-	_ = manager.Save(cred)
-	return newGitContext(*cred, s, logger)
-}
 
 type LocalRemote struct {
 	BaseURL       string
@@ -95,120 +48,3 @@ func (l LocalRemote) GetOrganization() string  { return l.Organization }
 func (l LocalRemote) GetEndpoint() string      { return l.BaseURL }
 func (l LocalRemote) GetBucketName() string    { return l.Bucket }
 func (l LocalRemote) GetStoragePrefix() string { return l.StoragePrefix }
-
-func (l LocalRemote) GetClient(remoteName string, logger *slog.Logger) (*GitContext, error) {
-	if username, password, err := gitrepo.GetRemoteBasicAuth(remoteName); err == nil && username != "" && password != "" {
-		l.BasicUsername = username
-		l.BasicPassword = password
-	}
-	projectID := l.GetProjectId()
-	bucketName := l.GetBucketName()
-	storagePrefix := l.GetStoragePrefix()
-	if strings.TrimSpace(l.GetOrganization()) != "" || strings.TrimSpace(bucketName) != "" || strings.TrimSpace(storagePrefix) != "" {
-		scope, err := gitrepo.ResolveBucketScope(
-			l.GetOrganization(),
-			projectID,
-			bucketName,
-			storagePrefix,
-		)
-		if err != nil {
-			return nil, err
-		}
-		bucketName = scope.Bucket
-		storagePrefix = scope.Prefix
-	}
-
-	cred := &syconf.Credential{APIEndpoint: l.BaseURL}
-	if l.BasicUsername != "" || l.BasicPassword != "" {
-		cred.KeyID = l.BasicUsername
-		cred.APIKey = l.BasicPassword
-	}
-
-	raw, err := syclient.New(l.BaseURL, syclient.WithBasicAuth(cred.KeyID, cred.APIKey))
-	if err != nil {
-		return nil, err
-	}
-	client, ok := raw.(*syclient.Client)
-	if !ok {
-		return nil, fmt.Errorf("unexpected syfon client type %T", raw)
-	}
-
-	return &GitContext{
-		Client:        client,
-		Organization:  l.GetOrganization(),
-		ProjectId:     projectID,
-		BucketName:    bucketName,
-		StoragePrefix: storagePrefix,
-		Logger:        logger,
-		Credential:    cred,
-	}, nil
-}
-
-func newGitContext(profileConfig syconf.Credential, remote Gen3Remote, logger *slog.Logger) (*GitContext, error) {
-	if _, err := url.Parse(profileConfig.APIEndpoint); err != nil {
-		return nil, err
-	}
-	projectID := remote.GetProjectId()
-	if projectID == "" {
-		return nil, fmt.Errorf("no gen3 project specified")
-	}
-
-	scope, err := gitrepo.ResolveBucketScope(
-		remote.GetOrganization(),
-		projectID,
-		remote.GetBucketName(),
-		remote.GetStoragePrefix(),
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	raw, err := syclient.New(profileConfig.APIEndpoint, syclient.WithBearerToken(profileConfig.AccessToken))
-	if err != nil {
-		return nil, err
-	}
-	client, ok := raw.(*syclient.Client)
-	if !ok {
-		return nil, fmt.Errorf("unexpected syfon client type %T", raw)
-	}
-
-	uploadConcurrency := int(gitrepo.GetGitConfigInt("lfs.concurrenttransfers", 4))
-	if uploadConcurrency < 1 {
-		uploadConcurrency = 1
-	}
-
-	return &GitContext{
-		Client:             client,
-		ProjectId:          projectID,
-		BucketName:         scope.Bucket,
-		Organization:       remote.GetOrganization(),
-		StoragePrefix:      scope.Prefix,
-		Upsert:             gitrepo.GetGitConfigBool("drs.upsert", false),
-		MultiPartThreshold: int64(gitrepo.GetGitConfigInt("drs.multipart-threshold", 5120)) * 1024 * 1024,
-		UploadConcurrency:  uploadConcurrency,
-		Logger:             logger,
-		Credential:         &profileConfig,
-	}, nil
-}
-
-func localRemoteFromGen3(gen3 *Gen3Remote, username string, password string) *LocalRemote {
-	return &LocalRemote{
-		BaseURL:       gen3.Endpoint,
-		ProjectID:     gen3.ProjectID,
-		Bucket:        gen3.Bucket,
-		Organization:  gen3.Organization,
-		StoragePrefix: gen3.StoragePrefix,
-		BasicUsername: strings.TrimSpace(username),
-		BasicPassword: strings.TrimSpace(password),
-	}
-}
-
-func WrapCredentialValidationError(remoteName string, err error) error {
-	if err == nil {
-		return nil
-	}
-	if strings.TrimSpace(remoteName) == "" {
-		return fmt.Errorf("%w. %s", err, credentialHelpSuffix)
-	}
-	return fmt.Errorf("%w. Remote %q requires refreshed credentials. %s", err, remoteName, credentialHelpSuffix)
-}
