+++ internal/drsdownload/bulk.go
@@ -1,10 +1,47 @@
-package drsremote
+package drsdownload
 
 import (
+	"context"
+	"fmt"
 	"strings"
 
+	"github.com/calypr/git-drs/internal/remoteruntime"
 	drsapi "github.com/calypr/syfon/apigen/client/drs"
 )
+
+func BulkAccessURLsForObjects(ctx context.Context, drsCtx *remoteruntime.GitContext, objects []drsapi.DrsObject) (map[string]drsapi.AccessURL, error) {
+	if drsCtx == nil || drsCtx.Client == nil {
+		return nil, fmt.Errorf("DRS client unavailable")
+	}
+	req, ok := bulkAccessRequest(objects)
+	if !ok {
+		return map[string]drsapi.AccessURL{}, nil
+	}
+
+	resp, err := drsCtx.Client.DRSAPI().GetBulkAccessURLWithResponse(ctx, req)
+	if err != nil {
+		return nil, err
+	}
+	if resp.JSON200 == nil {
+		return nil, fmt.Errorf("unexpected response: %d", resp.StatusCode())
+	}
+
+	out := map[string]drsapi.AccessURL{}
+	if resp.JSON200.ResolvedDrsObjectAccessUrls == nil {
+		return out, nil
+	}
+	for _, resolved := range *resp.JSON200.ResolvedDrsObjectAccessUrls {
+		if resolved.DrsObjectId == nil {
+			continue
+		}
+		objectID := *resolved.DrsObjectId
+		if strings.TrimSpace(objectID) == "" || strings.TrimSpace(resolved.Url) == "" {
+			continue
+		}
+		out[strings.TrimSpace(objectID)] = drsapi.AccessURL{Headers: resolved.Headers, Url: resolved.Url}
+	}
+	return out, nil
+}
 
 func bulkAccessRequest(objects []drsapi.DrsObject) (drsapi.BulkObjectAccessId, bool) {
 	req := drsapi.BulkObjectAccessId{}
