+++ internal/drslookup/scope.go
@@ -1,13 +1,26 @@
-package drsremote
+package drslookup
 
 import (
 	"fmt"
 	"strings"
 
-	drscommon "github.com/calypr/git-drs/internal/common"
 	drsapi "github.com/calypr/syfon/apigen/client/drs"
 	syfoncommon "github.com/calypr/syfon/common"
 )
+
+func ParseOrgProject(org, project string) (string, string) {
+	if org != "" {
+		return org, project
+	}
+	if project == "" {
+		return "", ""
+	}
+	if !strings.Contains(project, "-") {
+		return "default", project
+	}
+	parts := strings.SplitN(project, "-", 2)
+	return parts[0], parts[1]
+}
 
 func MatchesScope(obj *drsapi.DrsObject, organization, project string) bool {
 	return syfoncommon.DrsObjectMatchesScope(obj, organization, project)
@@ -18,7 +31,7 @@ func FindMatchingRecord(records []drsapi.DrsObject, organization, projectID stri
 		return nil, nil
 	}
 
-	org, project := drscommon.ParseOrgProject(strings.TrimSpace(organization), strings.TrimSpace(projectID))
+	org, project := ParseOrgProject(strings.TrimSpace(organization), strings.TrimSpace(projectID))
 	if org == "" {
 		return nil, fmt.Errorf("could not determine organization from inputs org=%q project=%q", organization, projectID)
 	}
