+++ internal/drspaths/paths.go
@@ -1,0 +1,12 @@
+package drspaths
+
+const (
+	LFSObjectsPath = ".git/lfs/objects"
+	DRSObjectsPath = ".git/drs/lfs/objects"
+	ConfigYAML     = "config.yaml"
+
+	DRSRefDir        = ".git/drs/objects"
+	DRSLogFile       = ".git/drs/git-drs.log"
+	ConfirmationYes  = "yes"
+	DRSDir           = ".git/drs"
+)
